requirements.txt
pandas
requests
pyyaml
azure-identity
msal
pyodbc
pytz

